Make json widget more configurable and consistent with other widgets
- Accept Width and height arguments, but default to the old hardcoded
values for backwards compability
- Apply "display: inline-block" style instead of "float: right". This will
wrap the element to the next line instead of overlapping the label if it
ends up too wide. This is consistent with the default behavior of form
elements like textarea which this widget replaces.
- Apply attrs to the container div to keep behavior consistent with
other standard widgets instead of silently ignoring the attrs argument.
- Allow all jsoneditor options to be configured with the "options"
argument
- The "mode" argument is not strictly needed anymore because the
mode can be as a part of 'options', but still accept it for backwards
compatibility
- Don't override the render method, just override the get_context
because we only need to set a few extra context vars. Keep behavior
consistent with other widgets, and play nicely with other apps like
"django-widget-tewaks"