Avoid HTML injection via unsafe JSON injection.
This change follows the Django docs' suggested approach for safely setting a JSON literal in a template: https://docs.djangoproject.com/en/3.2/ref/templates/builtins/#json-script.

In the old code, if a string literal within the JSON included '</script>', then that would be interpreted by the browser as closing the script tag, and thus allow arbitrary HTML to then be injected into the page. This is especially dangerous if the JSON for the widget can include user content, and if the widget is being used within the Django admin interface.

Fixes jmrivas86/django-json-widget#62.