Stop declaring support for Django versions before 2.2.
Version 2.1 has been out of extended support by Django since April 2019. Version 2.2 LTS is due to pass out of extended support this coming April. 3.0 and 3.1 passed out of extended support last year, but I've left them alone so that the supported versions still form a contiguous range.

My motivation is to drop support for version 2.1 and earlier is to to facilitate fixing an HTML injection using the 'json_script' template filter; see PR #64. This was introduced in version 2.1, but my hope is that it's reasonable to drop support for versions that have been unsupported by Django for years.

(I see that the version declared in requirements.txt is already excluding 2.1 and earlier; I'm unclear whether the difference is deliberate or not.)