Set the initial value of the textarea by script.
Since the value of the widget is now a Python dictionary, we can't use the string value for the textarea or we have a stringified dict, and not valid JSON

In practice, this meant that the widget could only save the content correctly if we'd run the onChange function, which sets the textarea content again.

With the new behaviour, I set the initial textarea content from script, so it's always set to a JSON string.

This doesn't degrade as gracefully for a client that doesn't run JavaScript, but since the textarea is set to 'display: none', the original behaviour didn't really degrade gracefully either.