Remove the claim of support for Django before version 3.2.
All of these older versions are now EOL, with 3.2 LTS being the oldest supported version by Django.

I'd like to add a claim of support for current Django version 4 onwards, but I think it makes sense to do that as a separate change to this one, and it deserves some work to ensure that it works as expected and in tests.